We have some initial numbers. Dunno which approach is better

+++ ComputeNoThreads.cs
@@ -2,7 +2,7 @@ public class ComputeNoThreads
 {
     public List<(int, int)> WeirdArray {get; set;}
 
-    public Int64 Compute()
+    public int Compute()
     {
         //a[0][0] * a[0][1] + a[1][0] * a[1][1] + ... + a[n][0] * a[n][1]
         var sum = 0;
